[feat]path변경 및 자장한 버그 수정
