+++ src/css/Main.css
@@ -165,6 +165,8 @@ footer {
   margin-top: 24px;
   font-size: 20px;
   font-family: "Pretendard-Bold";
+  height: 24PX;
+  overflow: hidden;
 }
 
 .card-explanation {
@@ -463,7 +465,7 @@ input {
   border: 1px solid #e1e1e1;
   border-radius: 10px;
   padding: 21px;
-  margin-bottom: 100px;
+  margin-bottom: 10px;
 }
 
 .studyDetail_review_top {
