+++ src/pages/StudyApplicants.jsx
@@ -4,7 +4,9 @@ import styles from '../css/ApplicantList.module.css'
 import ApplicantListUserName from '../components/ApplicantListUserName'
 import { authApi } from '../services/api'
 import { useLocation } from 'react-router'
-import StudyRejectReason from '../components/StudyRejectReason'
+import StudyRejectReason from '../components/Modal/StudyRejectReason'
+import UserName from '../components/UserName'
+import Profile from '../components/Modal/Profile'
 
 export default function ApplicantList() {
 
@@ -15,16 +17,44 @@ export default function ApplicantList() {
 
   // 정보 불러오기
   let applicantsList = [];
+  const [isApplicantsUpdate, setIsApplicantsUpldate] = useState(0);
 
-  useEffect(() => {
-    authApi(`studies/${studyId}/applicants`)
-      .then((response) => {
-        for (let i = 0; i < response.data.applicants.length; i++) {
-          applicantsList[i] = response.data.applicants[i]
-        }
-      })
-      .catch((e) => console.log(e))
-  })
+  // useEffect(() => {
+  //   authApi(`studies/${studyId}/applicants`)
+  //     .then((response) => {
+  //       for (let i = 0; i < response.data.applicants.length; i++) {
+  //         applicantsList[i] = response.data.applicants[i]
+  //       }
+  //     })
+  //     .catch((e) => console.log(e))
+  // }, [isApplicantsUpdate])
+
+  applicantsList = [{
+    "id": 1,
+    "nickname": "더미C",
+    "score": 100,
+    "applyDate": "2023-05-29T09:06:15.7453189"
+  }, {
+    "id": 2,
+    "nickname": "닉네임2",
+    "score": 92,
+    "applyDate": "2023-05-28T09:06:15.7453189"
+  }, {
+    "id": 3,
+    "nickname": "닉네임3",
+    "score": 93,
+    "applyDate": "2023-05-27T09:06:15.7453189"
+  }, {
+    "id": 4,
+    "nickname": "닉네임4",
+    "score": 98,
+    "applyDate": "2023-05-26T09:06:15.7453189"
+  }, {
+    "id": 5,
+    "nickname": "닉네임5",
+    "score": 95,
+    "applyDate": "2023-05-25T09:06:15.7453189"
+  }]
 
 
   // 승인 버튼
@@ -32,7 +62,7 @@ export default function ApplicantList() {
     authApi(`studies/${studyId}/applicants/${applierId}/approve`)
       .then((response) => {
         alert(`${nickName}님을 승인하였습니다`);
-        window.location.href = `${process.env.REACT_APP_BASE_URL}ApplicantList`;
+        setIsApplicantsUpldate(isApplicantsUpdate + 1)
       })
       .catch((e) => console.log(e));
   }
@@ -43,6 +73,8 @@ export default function ApplicantList() {
 
   const [input_reject_reason, setInput_reject_reason] = useState(false);
 
+  //프로필 모달
+  const [isview_profile_modal, setIsview_profile_modal] = useState(false);
 
   return (
     <div>
@@ -64,18 +96,18 @@ export default function ApplicantList() {
               </div>
 
 
-
-
-
               {
-                applicantsList.map((item, index) => {
+                applicantsList && applicantsList.map((item, index) => {
                   return (
                     <tr>
                       <div className={styles.list}>
 
                         <td className={styles.indextd}><div className={styles.data1}>{index + 1}</div></td>
 
-                        <td className={styles.nametd}><div className={styles.data2}><ApplicantListUserName userid={item.id} userName={item.nickname} /></div></td>
+                        <td className={styles.nametd}>
+                          <p onClick={() => setIsview_profile_modal(true)}>{item.nickname}</p>
+                          {isview_profile_modal && <Profile closeModal={setIsview_profile_modal} userId={item.id} />}
+                        </td>
 
                         <td className={styles.buttontd}><div className={styles.data3}>
                           <div className={styles.approve_button} onClick={() => handleStudyApprove(item.id, item.nickname)}>
